Introduce parameter length to Soundex
Standard indexes of length 4 may lead to a huge number of spellings with the same index (for a relatively small dictionary of 100,000 words it exceeds 100). Varying the length of a Soundex index allows for more specific sequences. This can be applied to longer words to reduce the number of spellings that share the same index.